% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for Package.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy of Predictions}
\usage{
calculate_accuracy(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A confusion matrix from binary classification predictions.}
}
\value{
Accuracy value.
}
\description{
Computes the accuracy from the confusion matrix for binary classification.
}
\examples{
predictions <- rbinom(100, 1, 0.5)
true_labels <- rbinom(100, 1, 0.5)
conf_matrix <- calculate_confusion_matrix(true_labels, predictions)
accuracy <- calculate_accuracy(conf_matrix)
}
