% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for Package.R
\name{calculate_confusion_matrix}
\alias{calculate_confusion_matrix}
\title{Calculate the Confusion Matrix}
\usage{
calculate_confusion_matrix(y, predictions)
}
\arguments{
\item{predictions}{Binary predictions from the model.}
}
\value{
A confusion matrix.
}
\description{
Computes the confusion matrix for binary classification predictions.
}
\details{
Title

 @param y True binary responses.
}
\examples{
true_values <- rbinom(100, 1, 0.5)
predicted_values <- rbinom(100, 1, 0.5)
conf_matrix <- calculate_confusion_matrix(true_values, predicted_values)
}
