% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for Package.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{Calculate Prevalence}
\usage{
calculate_prevalence(y)
}
\arguments{
\item{y}{Binary response vector.}
}
\value{
Prevalence value.
}
\description{
Computes the prevalence of the positive class in binary response data.
}
\examples{
response <- rbinom(100, 1, 0.7)
prevalence <- calculate_prevalence(response)
}
