% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Group8.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic Regression for Binary Classification}
\usage{
logistic_regression(x, y, B = 20, conf_level = 0.95)
}
\arguments{
\item{x}{A \code{matrix} or \code{data frame} of the x-variables (independent variables)
that will be used in your logistic regression.}

\item{y}{A \code{vector} containing your y (dependent) variable. This should be a
binary variable (two categories), and it must be a a \code{factor} to use in this
function.}

\item{B}{A \code{numeric} (integer) used to denote the number of bootstrap
simulations to perform in order to compute bootstrap confidence intervals for
the estimated coefficients.}

\item{conf_level}{A \code{numeric} used to specify the confidence level for
the bootstrap confidence intervals.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{coefficient_estimates}{Estimated values of the logistic regression coefficients}
     \item{bootstrap_intervals}{Boostrap confidence intervals for each of the estimated coefficients}
     \item{confusion_matrix}{A confusion matrix comparing the estimated classes based on your logistic regression model and the true classes}
     \item{confusion_metrics}{Metrics from your confusion matrix, including prevalence, accuracy, sensitivity, specificity, false discovery rate, and diagnostic odds ratio}
}
}
\description{
Calculate the estimated logistic regression coefficients for a binary
classification problem. The function also produces bootstrap confidence intervals
for each estimated coefficient. Additionally, the function will use the estimated
coefficients to produce predictions for the class for each data entry.
These predictions will be used to create a confusion matrix that compares the
estimated classes to the true classes.
}
\examples{
set.seed(123)
x <- data.frame(X1 = rnorm(20), X2 = rgamma(20, 5, 1), X3 = rbeta(20, 1, 1), X4 = sample(c('yes', 'no'), 20, TRUE, prob = c(0.7, 0.3)))
y <- as.factor(sample(c(0,1), 20, replace = TRUE, prob = c(0.3, 0.7)))
logistic_regression(x, y, B = 50, conf_level = 0.95)
}
\author{
John Skarmeas, Kwabena Bayity, Alireza Bidar
}
