% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for Package.R
\name{plot_logistic_curve}
\alias{plot_logistic_curve}
\title{Plot the Fitted Logistic Regression Curve}
\usage{
plot_logistic_curve(X, y, beta, predictor_index = 1)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of response variables (binary).}

\item{beta}{A vector of estimated coefficients from logistic regression.}

\item{predictor_index}{Index of the predictor to be used for plotting (default: 1).}
}
\description{
Plots the logistic regression curve based on fitted model coefficients.
}
\examples{
X <- matrix(rnorm(200), ncol = 2)
y <- rbinom(100, 1, 0.5)
beta <- logistic_regression(X, y)$coefficients
plot_logistic_curve(X, y, beta)
}
