% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for Package.R
\name{plot_selected_metrics_over_cutoffs}
\alias{plot_selected_metrics_over_cutoffs}
\title{Plot Selected Metrics Over Various Cutoff Values}
\usage{
plot_selected_metrics_over_cutoffs(fitted_probabilities, y, selected_metrics)
}
\arguments{
\item{fitted_probabilities}{Predicted probabilities from the logistic regression model.}

\item{y}{True binary responses.}

\item{selected_metrics}{A vector of selected metric names to plot.}
}
\description{
Plots selected metrics (e.g., accuracy, sensitivity) evaluated over a range of cutoff values.
}
\examples{
fitted_probs <- logistic_regression(X, y)$fitted_values
plot_selected_metrics_over_cutoffs(fitted_probs, y, c("Accuracy", "Sensitivity"))
}
